chore: formatting register-name tutorial for consistency
